feat: Promotion Service

+++ product/src/main/java/com/bkeuty/product/controller/admin/AdminCategoryController.java
@@ -1,7 +1,10 @@
 package com.bkeuty.product.controller.admin;
 
+import com.bkeuty.product.dto.auth.TokenValidationResponseDto;
 import com.bkeuty.product.entity.ProductCategory;
 import com.bkeuty.product.repository.ProductCategoryRepository;
+import com.bkeuty.product.service.authservice.AuthService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,23 +14,39 @@ import java.util.List;
 @RequestMapping("api/admin/category")
 public class AdminCategoryController {
     private final ProductCategoryRepository categoryRepository;
-
-    public AdminCategoryController(ProductCategoryRepository categoryRepository) {
+    private final AuthService authService;
+    public AdminCategoryController(ProductCategoryRepository categoryRepository, AuthService authService) {
         this.categoryRepository = categoryRepository;
+        this.authService = authService;
     }
 
     @GetMapping
-    public ResponseEntity<List<ProductCategory>> getAllCategories() {
+    public ResponseEntity<List<ProductCategory>> getAllCategories(@RequestHeader("Authorization") String bearerToken) {
+        TokenValidationResponseDto tokenValidationResponseDto = authService.validateToken(bearerToken);
+        if (tokenValidationResponseDto.getUserId() == null || tokenValidationResponseDto.getUserRole() == null
+                || !"admin".equals(tokenValidationResponseDto.getUserRole())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(categoryRepository.findAll());
     }
 
     @PostMapping
-    public ResponseEntity<ProductCategory> createCategory(@RequestBody ProductCategory category) {
+    public ResponseEntity<ProductCategory> createCategory(@RequestHeader("Authorization") String bearerToken,@RequestBody ProductCategory category) {
+        TokenValidationResponseDto tokenValidationResponseDto = authService.validateToken(bearerToken);
+        if (tokenValidationResponseDto.getUserId() == null || tokenValidationResponseDto.getUserRole() == null
+                || !"admin".equals(tokenValidationResponseDto.getUserRole())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(categoryRepository.save(category));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ProductCategory> updateCategory(@PathVariable Integer id, @RequestBody ProductCategory categoryDetails) {
+    public ResponseEntity<ProductCategory> updateCategory(@RequestHeader("Authorization") String bearerToken,@PathVariable Integer id, @RequestBody ProductCategory categoryDetails) {
+        TokenValidationResponseDto tokenValidationResponseDto = authService.validateToken(bearerToken);
+        if (tokenValidationResponseDto.getUserId() == null || tokenValidationResponseDto.getUserRole() == null
+                || !"admin".equals(tokenValidationResponseDto.getUserRole())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return categoryRepository.findById(id).map(category -> {
             category.setCategoryName(categoryDetails.getCategoryName());
             return ResponseEntity.ok(categoryRepository.save(category));
@@ -35,7 +54,12 @@ public class AdminCategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteCategory(@RequestHeader("Authorization") String bearerToken,@PathVariable Integer id) {
+        TokenValidationResponseDto tokenValidationResponseDto = authService.validateToken(bearerToken);
+        if (tokenValidationResponseDto.getUserId() == null || tokenValidationResponseDto.getUserRole() == null
+                || !"admin".equals(tokenValidationResponseDto.getUserRole())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return categoryRepository.findById(id).map(category -> {
             categoryRepository.delete(category);
             return ResponseEntity.ok().<Void>build();
